src/test/java/com/coekie/aoc2023/Day25Test.java
package com.coekie.aoc2023;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;

class Day25Test {
  private static final String example = """
      jqt: rhn xhk nvd
      rsh: frs pzl lsr
      xhk: hfx
      cmg: qnr nvd lhk bvb
      rhn: xhk bvb hfx
      bvb: xhk hfx
      pzl: lsr hfx nvd
      qnr: nvd
      ntq: jqt hfx bvb xhk
      nvd: lhk
      lsr: lhk
      rzs: qnr cmg lsr rsh
      frs: qnr lhk lsr""";

  @Test
  public void example() {
    assertEquals(54, Day25.solve(example));
  }

  @Test
  public void real() {
    assertEquals(525264, Day25.solve(InputReader.read("day25-input")));
  }
}